[Refactor] Move all speech recognizers into its own folder
