+++ app/input_items/recognizer_data.py
@@ -1,8 +1,8 @@
-from app.services.assembly_ai_speech_recognizer import AssemblyAiSpeechRecognizer
-from app.services.gladia_speech_recognizer import GladiaSpeechRecognizer
-from app.services.google_speech_recognizer import GoogleSpeechRecognizer
-from app.services.microsoft_azure_speech_recognizer import MicrosoftAzureSpeechRecognizer
-from app.services.open_ai_whisper_speech_recognizer import OpenAIWhisperSpeechRecognizer
+from app.services.speech_recognizers.assembly_ai_speech_recognizer import AssemblyAiSpeechRecognizer
+from app.services.speech_recognizers.gladia_speech_recognizer import GladiaSpeechRecognizer
+from app.services.speech_recognizers.google_speech_recognizer import GoogleSpeechRecognizer
+from app.services.speech_recognizers.microsoft_azure_speech_recognizer import MicrosoftAzureSpeechRecognizer
+from app.services.speech_recognizers.open_ai_whisper_speech_recognizer import OpenAIWhisperSpeechRecognizer
 
 
 class RecognizerData:
