+++ app/services/speech_recognizers/microsoft_azure_speech_recognizer.py
