+++ app/services/speech_recognizers/open_ai_whisper_speech_recognizer.py
