+++ tests/input_items/test_recognizer_data.py
@@ -1,10 +1,10 @@
 import pytest
 from app import RecognizerData
-from app.services.assembly_ai_speech_recognizer import AssemblyAiSpeechRecognizer
-from app.services.gladia_speech_recognizer import GladiaSpeechRecognizer
-from app.services.google_speech_recognizer import GoogleSpeechRecognizer
-from app.services.microsoft_azure_speech_recognizer import MicrosoftAzureSpeechRecognizer
-from app.services.open_ai_whisper_speech_recognizer import OpenAIWhisperSpeechRecognizer
+from app.services.speech_recognizers.assembly_ai_speech_recognizer import AssemblyAiSpeechRecognizer
+from app.services.speech_recognizers.gladia_speech_recognizer import GladiaSpeechRecognizer
+from app.services.speech_recognizers.google_speech_recognizer import GoogleSpeechRecognizer
+from app.services.speech_recognizers.microsoft_azure_speech_recognizer import MicrosoftAzureSpeechRecognizer
+from app.services.speech_recognizers.open_ai_whisper_speech_recognizer import OpenAIWhisperSpeechRecognizer
 
 
 class TestRecognizerData:
@@ -32,7 +32,7 @@ class TestRecognizerData:
         expected_str = (
             f"<class '{RecognizerData.__module__}.{RecognizerData.__name__}'>\n\n"
             'language_code = en-US\n'
-            "recognizer_class = <class 'app.services.google_speech_recognizer.GoogleSpeechRecognizer'>\n"
+            "recognizer_class = <class 'app.services.speech_recognizers.google_speech_recognizer.GoogleSpeechRecognizer'>\n"
             '_captions = False'
         )
         assert str(recognizer_data) == expected_str
