+++ tests/services/speech_recognizers/test_assembly_ai_speech_recognizer.py
