+++ tests/services/speech_recognizers/test_google_speech_recognizer.py
