+++ tests/services/speech_recognizers/test_microsoft_azure_speech_recognizer.py
