+++ tests/services/speech_recognizers/test_openai_whisper_speech_recognizer.py
